refactor: move service config
moving service config class into a extra file
